Re-organize FFI to aggregator and zkevm.
