+++ ffi/src/lib.rs
@@ -1,25 +1,5 @@
 #![feature(once_cell)]
 
-pub mod prove;
-pub mod verify;
-
-pub(crate) mod utils {
-    use std::{
-        ffi::{CStr, CString},
-        os::raw::c_char,
-    };
-
-    pub(crate) fn c_char_to_str(c: *const c_char) -> &'static str {
-        let cstr = unsafe { CStr::from_ptr(c) };
-        cstr.to_str().unwrap()
-    }
-
-    pub(crate) fn c_char_to_vec(c: *const c_char) -> Vec<u8> {
-        let cstr = unsafe { CStr::from_ptr(c) };
-        cstr.to_bytes().to_vec()
-    }
-
-    pub(crate) fn vec_to_c_char(bytes: Vec<u8>) -> *const c_char {
-        CString::new(bytes).unwrap().into_raw()
-    }
-}
+mod aggregator;
+mod utils;
+mod zkevm;
